moved db interactions to page folders
